pyproject.toml
[project]
name = "travel-assistant-livekit"
version = "0.1.0"
description = "LiveKit Voice Agent with LangGraph Multi-Model Orchestration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=3.2.0",
    "livekit>=1.0.20",
    "livekit-agents>=1.3.6",
    "livekit-plugins-openai>=1.3.6",
    "livekit-plugins-silero>=1.3.6",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
